[JIT] Add reference semantics to TorchScript classes
**Summary**
This commit adds reference semantics to TorchScript class types;
modifications made to them within TorchScript will be visible in Python.

**Test Plan**
This commit adds a unit test to `TestClassType` that checks that
modifications made to a class type instance passed into TorchScript are
visible in Python after executing the scripted function or module.

**Fixes**
This commit closes #41421.

ghstack-source-id: 28db27ea3fd8e29844d1bc327f41f2e173bd640e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44324