[qunat][graphmode][fx] Embedding/EmbeddingBag works in static quant qconfig
Summary:
When Embedding/EmbeddingBag are configured with static quant we'll skip inserting observer for
them in the graph and keep the op unchanged and print a warning.
This also aligns with eager mode behavior as well.

We'll enforce this behavior for other ops that only supports dynamic/weight_only quant but not static quant as well.

We used a global variable `DEFAULT_NOT_OBSERVED_QUANTIZE_HANDLER`, this is not exposed to user right now,
we can add that later if needed.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 6fef74bd681652c986918a9dce38d6678aa89dce
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48062